The "Art and Contracts" update
This update features a new UI with a sleek design. The buttons are intuitively placed to make the interface easy to traverse.
The other thing changed in this update is the "contract" table. It now contains the "nume" and "prenume" fields from the other two tables, and a reworked data creation and editing interface. You can now select from a list of "avocati" and "clienti" based on ID, "nume" and "prenume", so that you don't have to memorize anything. You just select the entries you want to be associated with your new contract, and you're good to go.
Other changes include various bug fixes.